+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [next]
 
+## [2.2.1] - 2025-01-21
+
 ### Added
 
 - Introduce LSP API Test Framework entry as `TestFrameworkType.Plugin.LSP`
@@ -1361,7 +1363,8 @@ The `2.0.0` release is completely rewritten. Please see [documentation page](htt
 
 - Support for attaching IntelliJ sources in IDEA
 
-[next]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.2.0...HEAD
+[next]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.2.1...HEAD
+[2.2.1]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.0.1...v2.1.0
 [2.0.1]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.0.0...v2.0.1